Update refactor code
(cherry picked from commit a225fdbe61ae7aee329fd3cf6985561049a6714b)
